@@ -1,12 +1,18 @@
 apply plugin: 'com.android.application'
 apply plugin: 'org.greenrobot.greendao' // apply plugin
 
+repositories {
+    maven {
+        url "https://linphone.org/maven_repository"
+    }
+}
+
 android {
-    compileSdkVersion 27
+    compileSdkVersion 29
     defaultConfig {
         applicationId "com.hunter.linphoneandroid"
-        minSdkVersion 18
-        targetSdkVersion 26
+        minSdkVersion 23
+        targetSdkVersion 29
         versionCode 201
         versionName "2.0.1"
     }
@@ -25,8 +31,8 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:recyclerview-v7:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
     // Glide
     implementation 'com.github.bumptech.glide:glide:4.8.0'
     annotationProcessor 'com.github.bumptech.glide:compiler:4.8.0'
@@ -49,11 +55,13 @@ dependencies {
     // Gson
     implementation 'com.google.code.gson:gson:2.8.5'
     // GreenDao
-    implementation 'org.greenrobot:greendao:3.2.2'
+    implementation 'org.greenrobot:greendao:3.3.0'
     // Logger
     implementation 'com.orhanobut:logger:2.2.0'
     // SystemBarTint
     implementation 'com.readystatesoftware.systembartint:systembartint:1.0.3'
     // AVLoadingIndicatorView
     implementation 'com.wang.avi:library:2.1.3'
+    // Liblinphone SDK
+    implementation 'org.linphone:linphone-sdk-android:4.5+'
 }
