@@ -1,13 +1,11 @@
 package com.hunter.linphoneandroid.base;
 
 import android.content.Context;
-import android.support.annotation.LayoutRes;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.LayoutRes;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-
-import com.hunter.linphoneandroid.base.RecyclerViewHolder;
 
 import java.util.List;
 
