@@ -1,6 +1,6 @@
 package com.hunter.linphoneandroid.base;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.SparseArray;
 import android.view.View;
 
