@@ -2,7 +2,7 @@ package com.hunter.linphoneandroid.ui.widget;
 
 import android.content.Context;
 import android.os.CountDownTimer;
-import android.support.v7.widget.AppCompatTextView;
+import androidx.appcompat.widget.AppCompatTextView;
 import android.util.AttributeSet;
 
 import com.hunter.fastandroid.R;
